fix: check exists before cleanup

+++ crates/craby_codegen/src/generators/cxx_generator.rs
@@ -542,22 +542,24 @@ impl Generator<CxxTemplate> for CxxGenerator {
     fn cleanup(ctx: &CodegenContext) -> Result<(), anyhow::Error> {
         let cxx_dir = cxx_dir(&ctx.root);
 
-        fs::read_dir(cxx_dir)?.try_for_each(|entry| -> Result<(), anyhow::Error> {
-            let path = entry?.path();
-            let file_name = path
-                .file_name()
-                .unwrap_or_default()
-                .to_string_lossy()
-                .to_string();
+        if cxx_dir.try_exists()? {
+            fs::read_dir(cxx_dir)?.try_for_each(|entry| -> Result<(), anyhow::Error> {
+                let path = entry?.path();
+                let file_name = path
+                    .file_name()
+                    .unwrap_or_default()
+                    .to_string_lossy()
+                    .to_string();
 
-            if file_name.starts_with("Cxx")
-                && (file_name.ends_with("Module.cpp") || file_name.ends_with("Module.hpp"))
-            {
-                fs::remove_file(&path)?;
-            }
+                if file_name.starts_with("Cxx")
+                    && (file_name.ends_with("Module.cpp") || file_name.ends_with("Module.hpp"))
+                {
+                    fs::remove_file(&path)?;
+                }
 
-            Ok(())
-        })?;
+                Ok(())
+            })?;
+        }
 
         Ok(())
     }
